src/requirements.txt
Arpeggio==1.5
achallonge-compat-fork==1.9.0
aiohttp==3.8.5
aiosignal==1.3.1
asyncio==3.4.3
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
boto3==1.28.50
botocore==1.31.50
capstone==4.0.2
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.4
contourpy==1.1.1
coverage==6.3.1
crypto==1.4.1
cycler==0.12.0
Flask==2.3.3
fonttools==4.43.0
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.5
loguru==0.7.0
MarkupSafe==2.1.3
matplotlib==3.8.0
multidict==6.0.4
Naked==0.1.31
nest-asyncio==1.5.8
nltk==3.6.7
numpy==1.22.3
packaging==23.2
pefile==2023.2.7
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
PyYAML==6.0
regex==2022.1.18
requests==2.28.1
s3transfer==0.6.2
seahorse==1.0.0
shellescape==3.8.1
shyaml==0.6.2
six==1.16.0
split==0.4
tqdm==4.62.3
urllib3==1.26.12
volatility3==1.0.0
Werkzeug==2.3.7
win32-setctime==1.1.0
yarl==1.9.2
